Updating loadlooks to work with new model and ImageFile.
loadlooks now works with the nested model structure as well as ImageFile field for images. Changed the schema in the load looks file. The new schema is in the new sheet in the same sheets link.

TODO(foobar): Make it work with multiple images for a single style. This will be done as a follow up.